+++ typescript/core/src/index.ts
@@ -2,8 +2,6 @@ export { McapIndexedReader } from "./McapIndexedReader";
 export { default as McapStreamReader } from "./McapStreamReader";
 export { McapWriter } from "./McapWriter";
 export type { McapWriterOptions } from "./McapWriter";
-export { McapRecordBuilder } from "./McapRecordBuilder";
-export { ChunkBuilder as McapChunkBuilder } from "./ChunkBuilder";
 export * as McapTypes from "./types";
 export * as McapConstants from "./constants";
 export type { IWritable } from "./IWritable";
