+++ typescript/core/src/internal/BufferBuilder.ts
